[bspwm/sxhkd] Stop X11 services on close
When shutting down bspwm, the various graphical services would crash
when it closed.  This adds a two way dependency on the services and a
stopping of the user target so that all of the services shutdown cleanly
when the WM closes rather than crash because there is no X11 session
running.