[cfgtools] Add priority for systempackages
The point of this is to establish ordering for different sources
allowing for fallbacks to be defined that won't get installed on systems
where a higher priority source exists (i.e. generic installers won't be
run if you can install the package/binary from the package manager).

The initial priority scheme is something like:
1: primary system package manager (pacman, apt, etc)
2: secondary system package manager (aur, deb files, etc)
3: language package managers (pip/virtualenv, cargo, etc)
10: remote repositories