[cfgtools/nix] Add NixPkgBin handler
This is another fallback type package installed that uses `nix-env` if
installed already on the system.  This will install the binary into the
active nix-profile, but then also symlink directly to the binary outside
the profile to avoid profile changing causing churn.

Currently untested, will be validating it in the future.