[vim] Manually start LSPs
Because, for some reason, initial ftplugin loading doesn't trigger the
autostart for LSP servers, manually start them after setup.  This seems
to be due to the order things are running with setup and start, if you
load another file of that type, it launches.  Don't want to move these
as there are language specific fixes that are required (like
injecting the python virtualenv for python files)