[legacy] Move over xmonad legacy configs
I am trying to move some/most of the legacy (read: unused) configs I
originally had for posterity/potential re-use in the future.  Xmonad was
one of the bigger ones.

This attempts to translate some of the various concepts into the new
framework but doesn't go about approaching the main code (all the
haskell config) as it would be untested.