[utils/cgroups] Initial cgroups setup
Systemd provides some utility when utilizing the cgroups/slice system
for running programs in scopes.  This is just some configuration to run
programs in groups (and a wrapper to make it easier to auto categorize
apps).  Eventually need to better document/tool setting limits and
probably running something like `systemd-oomd` to utilize some of the
scoping advantages.