[wayland/ags] Update config to be system configurable
This adds the idea of the setup to be configurable via a "settings.json"
file that will allow for systems to override default settings.  Then
declare the default settings.  Then define the settings setup via a
helper File type to ensure the destination setting.  Also need to
gitignore the settings file since it technically gets symlinked into the
repo (since we lazily symlink the whole config directory into the XDG
folder).

Changes to the config file around this include:
- Make the module imports dynamic and defined via the settings
- Move the accent creation into a widget
- Make the accent optional via setting
- Define monitor via setting, update accent and notifications to honor
  monitor setting (may need to revisit if doing multiple bars?)