[ags] Split out popup logic to helper widget
This pulls the drawing/managing logic for popups from the bluetooth
module (which is the only one with a real popup) into a consolidated
helper so that I can start defining popups for other modules.  There is
still some tweaking for UX like timeout refreshes if interacted with or
multiple popups happening.