[wayland/ags] Add mpris active override
This adds a right click menu to the mpris widget to allow overriding the
active player.  This is kind of a hack where it just bumps the timestamp
to take precedence.  If another player updates after, this goes away.

Some TODO around cleaning up/making the selector a bit more informative,
but starting with the basics for now.