[systems/hershel] Add system specific ags module
This some small tweaks to AGS to enable a special system module for my
desktop where it adds a button to the sidebar to shift which monitor
notifications show up on.  This is pokes holes in the notification logic
to control the monitor, adds the module to the system settings, and
links the special system module into place (and ignores it).