[utils/diskmgr] Add new disk manager script
This is a python script to try and simplify and manage mounting and
unmounting my various sd cards.  Right now the script is just setup to
generate fstab entries for these, mount, and unmount them.  In the
future I want to expand this some for setting up new cards, hooking into
udev in some way to prompt to mount stuff, and possibly some other
scenario handling.