[wayland/hyprland] Handle workspace switching with script
The main purpose is to have workspace switching with multiple monitors
work how I expect.  This is mainly already done with the `m~#` handler,
but that fails to work if the workspace doesn't exist in the list (i.e.
it goes away when empty).  This script just calculates the workspace
number based on the active monitor and uses an offset, which achieves
what I expect.

May think about modifying this to ensure any created workspace gets the
correct properties, but can tackle that in the future.